+++ docker/.env
@@ -14,4 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-gevent==1.4.0
+COMPOSE_PROJECT_NAME=superset
+POSTGRES_DB=superset
+POSTGRES_HOST=postgres
+POSTGRES_PASSWORD=superset
+POSTGRES_PORT=5432
+POSTGRES_USER=superset
+# Add the mapped in /app/pythonpath_docker which allows devs to override stuff
+PYTHONPATH=/app/pythonpath:/app/pythonpath_docker
+REDIS_HOST=redis
+REDIS_PORT=6379
+
+FLASK_ENV=development
+SUPERSET_ENV=development
+SUPERSET_LOAD_EXAMPLES=yes
