+++ docker/docker-entrypoint.sh
@@ -15,24 +15,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-set -ex
+set -eo pipefail
 
-if [ "$#" -ne 0 ]; then
-    exec "$@"
-elif [ "$SUPERSET_ENV" = "development" ]; then
-    celery worker --app=superset.sql_lab:celery_app --pool=gevent -Ofair &
-    # needed by superset runserver
-    (cd superset/assets/ && npm ci)
-    (cd superset/assets/ && npm run dev) &
-    FLASK_ENV=development FLASK_APP="superset.app:create_app()" flask run -p 8088 --with-threads --reload --debugger --host=0.0.0.0
-elif [ "$SUPERSET_ENV" = "production" ]; then
-    celery worker --app=superset.sql_lab:celery_app --pool=gevent -Ofair &
-    exec gunicorn --bind  0.0.0.0:8088 \
-        --workers $((2 * $(getconf _NPROCESSORS_ONLN) + 1)) \
+if [ "${#}" -ne 0 ]; then
+    exec "${@}"
+else
+    gunicorn --bind  "0.0.0.0:${SUPERSET_PORT}" \
+        --workers 1 \
         --timeout 60 \
         --limit-request-line 0 \
         --limit-request-field_size 0 \
-        "superset.app:create_app()"
-else
-    superset --help
+        "${FLASK_APP}"
 fi
