+++ docker/pythonpath_dev/.gitignore
@@ -1,4 +1,3 @@
-#!/usr/bin/env bash
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
@@ -15,19 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-set -ex
 
-# Create an admin user (you will be prompted to set username, first and last name before setting a password)
-export FLASK_APP="superset.app:create_app()"
-flask fab create-admin
-
-# Initialize the database
-superset db upgrade
-
-if [ "$SUPERSET_LOAD_EXAMPLES" = "yes" ]; then
-    # Load some data to play with
-    superset load_examples
-fi
-
-# Create default roles and permissions
-superset init
+# Ignore everything
+*
+# DON'T ignore the .gitignore
+!.gitignore
+!superset_config.py
+!superset_config_local.example
